Move Comments To Its Own `/comment` Route
