+++ request.rest
@@ -35,11 +35,11 @@ Content-Type: application/json
 
 ###
 
-POST http://localhost:3000/post/comment
+POST http://localhost:3000/comment
 Content-Type: application/json
 
 {
-    "postId": "674cb3e4e0429dd40c4ca29b",
+    "postId": "674cb68f8dd29399c22f5f65",
     "sender": "Tal",
     "content": "HELLO COMMENT!!"
 }
