+++ routes/comments_routes.js
@@ -1,0 +1,8 @@
+
+const express = require('express');
+const router = express.Router();
+const Comment = require('../controllers/comments_controller');
+
+router.post('/', Comment.createComment);
+
+module.exports = router;
