+++ routes/posts_routes.js
@@ -11,6 +11,4 @@ router.get('/:id', Post.getPostById);
 
 router.put('/:id', Post.updatePostById);
 
-router.post('/comment', Post.createComment);
-
 module.exports = router;
